renamed health-routes -> status-routes
