+++ express/app.ts
@@ -1,11 +1,11 @@
 import express from "express";
 import serverless from "serverless-http";
-import healthCheckRoutes from "./routes/health-routes";
+import statusCheckRoutes from "./routes/status-routes";
 import usersRouter from "./routes/users/users";
 
 const app = express();
 
-app.use("/health", healthCheckRoutes);
+app.use("/health", statusCheckRoutes);
 app.use("/users", usersRouter);
 
 app.use((req, res, next) => {
